// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/ReviewCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ReviewResult } from './ReviewResult';
import {
    ReviewResultFromJSON,
    ReviewResultFromJSONTyped,
    ReviewResultToJSON,
    ReviewResultToJSONTyped,
} from './ReviewResult';

/**
 * 
 * @export
 * @interface ReviewCreate
 */
export interface ReviewCreate {
    /**
     * 
     * @type {Date}
     * @memberof ReviewCreate
     */
    reviewTime: Date;
    /**
     * 
     * @type {ReviewResult}
     * @memberof ReviewCreate
     */
    result: ReviewResult;
    /**
     * 
     * @type {number}
     * @memberof ReviewCreate
     */
    durationSeconds?: number | null;
    /**
     * 
     * @type {string}
     * @memberof ReviewCreate
     */
    studentAnswer?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ReviewCreate
     */
    remark?: string | null;
}



/**
 * Check if a given object implements the ReviewCreate interface.
 */
export function instanceOfReviewCreate(value: object): value is ReviewCreate {
    if (!('reviewTime' in value) || value['reviewTime'] === undefined) return false;
    if (!('result' in value) || value['result'] === undefined) return false;
    return true;
}

export function ReviewCreateFromJSON(json: any): ReviewCreate {
    return ReviewCreateFromJSONTyped(json, false);
}

export function ReviewCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReviewCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'reviewTime': (new Date(json['reviewTime'])),
        'result': ReviewResultFromJSON(json['result']),
        'durationSeconds': json['durationSeconds'] == null ? undefined : json['durationSeconds'],
        'studentAnswer': json['studentAnswer'] == null ? undefined : json['studentAnswer'],
        'remark': json['remark'] == null ? undefined : json['remark'],
    };
}

export function ReviewCreateToJSON(json: any): ReviewCreate {
    return ReviewCreateToJSONTyped(json, false);
}

export function ReviewCreateToJSONTyped(value?: ReviewCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'reviewTime': ((value['reviewTime']).toISOString()),
        'result': ReviewResultToJSON(value['result']),
        'durationSeconds': value['durationSeconds'],
        'studentAnswer': value['studentAnswer'],
        'remark': value['remark'],
    };
}