// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/WrongUpdate.ts
export interface WrongUpdate {
    /**
     * 
     * @type {string}
     * @memberof WrongUpdate
     */
    studentId: string;
    /**
     * 
     * @type {string}
     * @memberof WrongUpdate
     */
    subjectId: string;
    /**
     * 
     * @type {WrongSource}
     * @memberof WrongUpdate
     */
    sourceType: WrongSource;
    /**
     * 题型动态字典 code，来自 question_type
     * @type {string}
     * @memberof WrongUpdate
     */
    questionType?: string | null;
    /**
     * 
     * @type {string}
     * @memberof WrongUpdate
     */
    questionText: string;
    /**
     * 
     * @type {string}
     * @memberof WrongUpdate
     */
    studentAnswer?: string;
    /**
     * 
     * @type {string}
     * @memberof WrongUpdate
     */
    correctAnswer?: string;
    /**
     * 
     * @type {string}
     * @memberof WrongUpdate
     */
    analysisText?: string;
    /**
     * 
     * @type {string}
     * @memberof WrongUpdate
     */
    errorType?: string;
    /**
     * 
     * @type {number}
     * @memberof WrongUpdate
     */
    difficulty?: number;
    /**
     * 
     * @type {Array<KnowledgeLink>}
     * @memberof WrongUpdate
     */
    knowledgePoints?: Array<KnowledgeLink>;
    /**
     * 
     * @type {number}
     * @memberof WrongUpdate
     */
    version: number;
    /**
     * 
     * @type {WrongStatus}
     * @memberof WrongUpdate
     */
    status?: WrongStatus;
}