// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/ReviewResult.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ReviewResult = {
    Correct: 'CORRECT',
    Partial: 'PARTIAL',
    Wrong: 'WRONG',
    Unknown: 'UNKNOWN'
} as const;
export type ReviewResult = typeof ReviewResult[keyof typeof ReviewResult];


export function instanceOfReviewResult(value: any): boolean {
    for (const key in ReviewResult) {
        if (Object.prototype.hasOwnProperty.call(ReviewResult, key)) {
            if (ReviewResult[key as keyof typeof ReviewResult] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ReviewResultFromJSON(json: any): ReviewResult {
    return ReviewResultFromJSONTyped(json, false);
}

export function ReviewResultFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReviewResult {
    return json as ReviewResult;
}

export function ReviewResultToJSON(value?: ReviewResult | null): any {
    return value as any;
}

export function ReviewResultToJSONTyped(value: any, ignoreDiscriminator: boolean): ReviewResult {
    return value as ReviewResult;
}