// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/WrongSource.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const WrongSource = {
    Practice: 'PRACTICE',
    Exam: 'EXAM'
} as const;
export type WrongSource = typeof WrongSource[keyof typeof WrongSource];


export function instanceOfWrongSource(value: any): boolean {
    for (const key in WrongSource) {
        if (Object.prototype.hasOwnProperty.call(WrongSource, key)) {
            if (WrongSource[key as keyof typeof WrongSource] === value) {
                return true;
            }
        }
    }
    return false;
}

export function WrongSourceFromJSON(json: any): WrongSource {
    return WrongSourceFromJSONTyped(json, false);
}

export function WrongSourceFromJSONTyped(json: any, ignoreDiscriminator: boolean): WrongSource {
    return json as WrongSource;
}

export function WrongSourceToJSON(value?: WrongSource | null): any {
    return value as any;
}

export function WrongSourceToJSONTyped(value: any, ignoreDiscriminator: boolean): WrongSource {
    return value as WrongSource;
}